refactor: Add StarSystem entity to remove data redundancy
Before this commit, each row in the planet database table would
hold its host star name among other information pertaining to
the star system it resides in. This data is duplicated for planets
that share the same system which is not ideal. By introducing
the StarSystem database table, this removes this problem.